Notion Release Notes Agent/requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
slack-sdk>=3.23.0
scalekit-sdk-python>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
